+++ packages/fastboard-core/full.d.ts
@@ -1,1 +1,6 @@
+declare global {
+  interface Window {
+    global: any;
+  }
+}
 export * from "./dist/full";