+++ packages/fastboard-core/src/full.ts
@@ -1,2 +1,5 @@
+if (typeof window !== "undefined") {
+  (window as any).global = (window as any).global || {};
+}
 export * from "./lite";
 export * from "./behaviors";