+++ packages/fastboard-ui/src/components/Toolbar/components/helper.ts
@@ -23,7 +23,6 @@ export function tooltip(text: string, hotkey?: HotKey) {
   return outer;
 }
 
-
 export function rgbToHex(r: number, g: number, b: number) {
   const hex = ((r << 16) + (g << 8) + b).toString(16).padStart(6, "0");
   return "#" + hex;